rename base -> data_file_manager to make it more obvious what's in the file
