+++ pyqcrbox/pyqcrbox/data_management/__init__.py
@@ -1,4 +1,4 @@
-from .base import DataFileManager, DatasetNotFoundError
 from .data_file import Dataset, DatasetResponse
+from .data_file_manager import DataFileManager, DatasetNotFoundError
 from .dummy_data_file_manager import DummyDataFileManager
 from .nats_data_file_manager import NatsDataFileManager
